+++ ZaloPay/ui.widget/src/main/java/com/zalopay/ui/widget/password/indicator/LoadingIndicatorView.java
@@ -81,7 +81,7 @@ public class LoadingIndicatorView extends View {
     @Override
     protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
         int parentWidth = MeasureSpec.getSize(widthMeasureSpec);
-        DEFAULT_SIZE = (int) (parentWidth / (DEFAULT_SCREEN / 2));
+        DEFAULT_SIZE = parentWidth / (DEFAULT_SCREEN / 2);
         int width = measureDimension(dp2px(DEFAULT_SIZE), widthMeasureSpec);
         int height = measureDimension(dp2px(DEFAULT_SIZE), heightMeasureSpec);
         setMeasuredDimension(width, height);
